Move ADIOS2 specific switchTypes to ADIOS2Auxiliary.hpp
Also, add a missing newline
