+++ src/IO/ADIOS/ADIOS2PreloadAttributes.cpp
@@ -25,9 +25,6 @@
 #include "openPMD/IO/ADIOS/ADIOS2PreloadAttributes.hpp"
 
 #include "openPMD/Datatype.hpp"
-#define OPENPMD_SWITCHTYPE_ADIOS2
-#include "openPMD/DatatypeHelpers.hpp"
-#undef OPENPMD_SWITCHTYPE_ADIOS2
 #include "openPMD/IO/ADIOS/ADIOS2Auxiliary.hpp"
 #include "openPMD/auxiliary/StringManip.hpp"
